chore(utils): seperates interfaces from the core module
seperated the interfaces to a seperate file to make the code more cleaner
