+++ packages/utils/modify-config-helper.ts
@@ -5,23 +5,8 @@ import * as path from "path";
 import * as yeoman from "yeoman-environment";
 import * as Generator from "yeoman-generator";
 
+import { TransformConfig } from "./types/Config";
 import runTransform from "./scaffold";
-
-export interface Config extends Object {
-	item?: {
-		name: string;
-	};
-	topScope?: string[];
-	configName?: string;
-	merge: object;
-	webpackOptions: object;
-}
-
-export interface TransformConfig extends Object {
-	configPath?: string;
-	configFile?: string;
-	config?: Config;
-}
 
 const DEFAULT_WEBPACK_CONFIG_FILENAME = "webpack.config.js";
