+++ packages/utils/types/Config.ts
@@ -1,0 +1,15 @@
+export interface Config extends Object {
+	item?: {
+		name: string;
+	};
+	topScope?: string[];
+	configName?: string;
+	merge: object;
+	webpackOptions: object;
+}
+
+export interface TransformConfig extends Object {
+	configPath?: string;
+	configFile?: string;
+	config?: Config;
+}
